create transfer api

+++ db/sqlc/tranfer.sql.go
@@ -18,7 +18,7 @@ RETURNING id, from_account_id, to_account_id, amount, created_at
 type CreateTransferParams struct {
 	FromAccountID int64 `json:"from_account_id"`
 	ToAccountID   int64 `json:"to_account_id"`
-	Amount        int32 `json:"amount"`
+	Amount        int64 `json:"amount"`
 }
 
 func (q *Queries) CreateTransfer(ctx context.Context, arg CreateTransferParams) (Tranfer, error) {
